修改了编辑开单细节模块对sql空调用的bug

+++ management/src/main/java/org/example/management/service/impl/SlipDetailServiceImpl.java
@@ -30,26 +30,47 @@ public class SlipDetailServiceImpl implements SlipDetailService {
     @Override
     public void update(Integer slipId,List<SlipDetail> slipDetails) {
         List<SlipDetail> ori = slipDetailMapper.list(slipId,null);
-        List<SlipDetail> addList  = new ArrayList<>();
-        List<SlipDetail> updateList = new ArrayList<>();
         List<SlipDetail> deleteList = new ArrayList<>(ori);
-        for(SlipDetail sp : slipDetails){
-            if(sp.getId()==null){
-                addList.add(sp);
-            }else{
-                updateList.add(sp);
-            }
-        }
-        slipDetailMapper.insert(addList);
-        slipDetailMapper.update(updateList);
         List<Integer> deleteIds = new ArrayList<>();
-        for(SlipDetail sp : deleteList){
-            deleteIds.add(sp.getId());
+        if(slipDetails == null || slipDetails.isEmpty()){
+            if(deleteList != null && !deleteList.isEmpty()){
+                for(SlipDetail sp : deleteList){
+                    deleteIds.add(sp.getId());
+                }
+                slipDetailMapper.delete(deleteIds);
+                log.info("删除细节：{}",deleteList);
+            }
+        }else{
+            List<SlipDetail> addList  = new ArrayList<>();
+            List<SlipDetail> updateList = new ArrayList<>();
+            for(SlipDetail sp : slipDetails){
+                if(sp.getId()==null){
+                    addList.add(sp);
+                }else{
+                    updateList.add(sp);
+                }
+            }
+            if (addList != null && !addList.isEmpty()){
+                slipDetailMapper.insert(addList);
+                log.info("插入新细节：{}",addList);
+            }
+            if (updateList != null && !updateList.isEmpty()){
+                slipDetailMapper.update(updateList);
+                log.info("更新新细节：{}",updateList);
+            }
+
+            for(SlipDetail sp : deleteList){
+                deleteIds.add(sp.getId());
+            }
+            for(SlipDetail sp :updateList){
+                deleteIds.remove(sp.getId());
+            }
+            if(deleteIds != null && !deleteIds.isEmpty()){
+                slipDetailMapper.delete(deleteIds);
+            }
+
         }
-        for(SlipDetail sp :updateList){
-            deleteIds.remove(sp.getId());
-        }
-        slipDetailMapper.delete(deleteIds);
+
     }
 
     @Override
